Extract to_json() to utils (reutilization)
